+++ frontend/src/components/UserList.tsx
@@ -19,6 +19,10 @@ const UserList: React.FC<Props> = ({
   const [loading, setLoading] = useState<boolean>(true);
   const [error, setError] = useState<string | null>(null);
 
+  const [selectedUserId, setSelectedUserId] = useState<number | null>(null); // 선택된 유저 ID
+  const [selectedUserName, setSelectedUserName] = useState<string>(""); // 선택된 유저 이름
+  const [showCreateUI, setShowCreateUI] = useState<boolean>(false); // 채팅방 생성 UI 표시 여부
+
   useEffect(() => {
     const fetchUsers = async () => {
       try {
@@ -49,7 +53,50 @@ const UserList: React.FC<Props> = ({
     fetchUsers();
   }, [managementDashboardName]);
 
-  const createChatRoom = async (userId: number) => {
+  const checkChatRoomExistence = async (userId: number): Promise<boolean> => {
+    try {
+      const response = await fetch(
+        `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/v1/chats/chatRooms/exists?userId=${userId}&chatRoomType=ONE_TO_ONE`,
+        {
+          method: "GET",
+          headers: {
+            "Content-Type": "application/json",
+          },
+          credentials: "include",
+        }
+      );
+
+      if (!response.ok) {
+        throw new Error("채팅방 존재 여부 확인 중 오류가 발생했습니다.");
+      }
+
+      const data = await response.json();
+      return data.data; // true: 채팅방 존재, false: 채팅방 없음
+    } catch (err) {
+      console.error("채팅방 존재 여부 확인 실패:", err);
+      return false;
+    }
+  };
+
+  const handleChatRoomCheck = async (userId: number, userName: string) => {
+    const exists = await checkChatRoomExistence(userId);
+
+    if (exists) {
+      alert("이미 존재하는 채팅방입니다.");
+      setShowCreateUI(false); // 채팅방 생성 UI 숨기기
+    } else {
+      setSelectedUserId(userId); // 선택된 유저 ID 설정
+      setSelectedUserName(userName); // 선택된 유저 이름 설정
+      setShowCreateUI(true); // 채팅방 생성 UI 표시
+    }
+  };
+
+  const createChatRoom = async () => {
+    if (!selectedUserId || !selectedUserName) {
+      alert("채팅방 생성에 필요한 정보가 부족합니다.");
+      return;
+    }
+
     try {
       const response = await fetch(
         `${process.env.NEXT_PUBLIC_API_BASE_URL}/api/v1/chats/chatRooms`,
@@ -60,9 +107,9 @@ const UserList: React.FC<Props> = ({
           },
           credentials: "include",
           body: JSON.stringify({
-            userId, // 선택된 유저 ID
-            roomName: "1:1 채팅방", // 채팅방 이름
-            roomType: "ONE_TO_ONE", // 채팅방 유형
+            userId: selectedUserId, // 선택된 유저 ID
+            roomName: selectedUserName, // 선택된 유저 이름을 채팅방 이름으로 사용
+            roomType: "ONE_TO_ONE", // 채팅방 유형 고정
           }),
         }
       );
@@ -72,6 +119,10 @@ const UserList: React.FC<Props> = ({
       }
 
       alert("채팅방이 성공적으로 생성되었습니다!");
+      window.location.reload(); // 새로고침
+      setSelectedUserId(null); // 선택된 유저 초기화
+      setSelectedUserName(""); // 선택된 유저 이름 초기화
+      setShowCreateUI(false); // 채팅방 생성 UI 숨기기
     } catch (err) {
       alert("채팅방 생성 중 오류가 발생했습니다.");
     }
@@ -95,13 +146,41 @@ const UserList: React.FC<Props> = ({
             </div>
             <button
               className="bg-blue-500 text-white px-4 py-2 rounded"
-              onClick={() => createChatRoom(user.id)} // 채팅방 생성 요청
+              onClick={() => handleChatRoomCheck(user.id, user.name)} // 채팅방 존재 여부 확인
             >
               채팅방 생성
             </button>
           </li>
         ))}
       </ul>
+
+      {showCreateUI && (
+        <div
+          className="fixed inset-0 flex items-center justify-center bg-black bg-opacity-50"
+          style={{ zIndex: 1000 }}
+        >
+          <div className="bg-white p-6 rounded shadow-lg w-96">
+            <h3 className="text-lg font-bold mb-2">채팅방 생성</h3>
+            <p className="text-sm mb-2">
+              선택된 유저: <strong>{selectedUserName}</strong>
+            </p>
+            <div className="flex justify-end gap-2">
+              <button
+                className="bg-gray-500 text-white px-4 py-2 rounded"
+                onClick={() => setShowCreateUI(false)} // 창 닫기
+              >
+                취소
+              </button>
+              <button
+                className="bg-green-500 text-white px-4 py-2 rounded"
+                onClick={createChatRoom} // 채팅방 생성 요청
+              >
+                생성
+              </button>
+            </div>
+          </div>
+        </div>
+      )}
     </div>
   );
 };
